Move user xattr tests into their own package and organise by functionality
